HTTPClient,Serializer,Computer
Now we can connect with HTTPClient witn class ServerRespons by using method get and post. We can deserialize data from server to list of Computers. We have class Computer which field are coresponding to JSON data from server. (Usage of get method is in ListOfCompiterDataFragement file in onResume and getComputers).